Skip certificate verification
According to the documentation of Netdata, if you use self-signed certificates, as we do in our setup, you need to skip the verification of the certificate. The traffic remains encrypted and you still need to have the parent node's certificate in order to establish the connection.

Reference: https://learn.netdata.cloud/docs/deployment-in-production/streaming-and-replication/streaming-configuration-reference#streamconf